requirements.txt
nextdraw-api @ https://software-download.bantamtools.com/nd/api/nextdraw_api.zip
lxml~=5.3.0